refactor(influxdb): anomalous datapoints are now duplicated into a seperate measurement instead of being set as a tag
Issue: Implement RePAD2 #5

Description:
This change was done because events in InfluxDB are not modifiable. Therefore, anomalous datapooints are now duplicated to a seperate __measurement in InfluxDB instead of using the tag "is_anomaly". This makes the process of "tagging" anomalies a lot simpler as the detected anomalous datapoints are simply duplicated and written to InfluxDB